Renaming compiler's logger to avoid confusion or overloads
