+++ src/template/compiler/compiler_logger.js
